requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-anthropic>=0.1.0
llamaindex>=0.9.0
streamlit>=1.30.0
anthropic>=0.18.0

# Vector store
chromadb>=0.4.0
faiss-cpu>=1.7.0

# Embeddings
sentence-transformers>=2.2.0

# Web search
tavily-python>=0.3.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
